worker/requirements.txt
fastapi==0.116.0
pydantic==2.11.7
python-dotenv==1.1.1
spotipy==2.25.1
typer==0.16.0
uvicorn==0.35.0
yt_dlp==2025.6.30
